refactor(infra): extract MockScaffoldingResponder from ProviderFactory
